backend/requirements.txt
# ── Core Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ── Pydantic / Settings ────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.0

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
aiosqlite==0.20.0
asyncpg==0.29.0          # PostgreSQL async driver (prod)

# ── AI / LLM / Embeddings ──────────────────────────────────────────────────────
openai==1.35.3
tiktoken==0.7.0

# ── Vector Database ────────────────────────────────────────────────────────────
faiss-cpu==1.8.0
numpy==1.26.4

# ── Caching / Task Queue ───────────────────────────────────────────────────────
redis[asyncio]==5.0.4
celery==5.4.0
flower==2.0.1             # Celery monitoring dashboard

# ── Logging ────────────────────────────────────────────────────────────────────
structlog==24.2.0

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
pytest-cov==5.0.0

# ── Dev / Code Quality ─────────────────────────────────────────────────────────
ruff==0.4.9
mypy==1.10.0
